fix some problem by review advices
